requirements.txt
zwift-client==0.2.0
protobuf==3.20.3
paho-mqtt==2.1.0
numpy==2.0.0